+++ Library/Stubs/Generator.php
@@ -23,7 +23,6 @@ use Zephir\CompilerFile;
 use Zephir\Exception;
 
 use function array_key_exists;
-use function count;
 use function in_array;
 
 /**
@@ -313,7 +312,7 @@ class Generator
         if ($method->hasReturnTypes()) {
             $supported = 0;
 
-            if (array_key_exists('object', $method->getReturnTypes()) && 1 === count($method->getReturnClassTypes())) {
+            if (array_key_exists('object', $method->getReturnTypes())) {
                 $return = key($method->getReturnClassTypes());
                 ++$supported;
             }
@@ -348,7 +347,7 @@ class Generator
             }
 
             // PHP doesn't support multiple return types (yet?)
-            if ($supported > 1) {
+            if ($supported > 1 || array_key_exists('variable', $method->getReturnTypes())) {
                 $return = '';
             }
         } elseif ($method->isVoid()) {