+++ Library/Variable.php
@@ -219,7 +219,7 @@ class Variable implements TypeAwareInterface
      *
      * @return string
      */
-    public function getType()
+    public function getType(): string
     {
         return $this->type;
     }